Abandon legacy branch
There was a `legacy` branch where golden images prescribed what Ruby
versions would be shipped, and downstream consumers could just select
from them. That sorta goes against the spirit of this original repo.

The only supported consumer of this repo right now is kingdonb/ob-mirror

It selects the `ruby-3.0.7` Ruby version, because the [Beeminder gem][]
demands it, due to the selected version of ActiveSupport.

(We'll try to help them upgrade, so we don't have to be on that old,
unsupported version of Ruby anymore! We should be on at least 3.1.6,
with plans to upgrade to 3.2 before March...)

[Beeminder gem]: https://github.com/beeminder/beeminder-gem

Signed-off-by: Kingdon B <kingdon@urmanac.com>